requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.10.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
scipy>=1.10.0

# Biological NLP
scispacy>=0.5.1
en-core-sci-lg @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_core_sci_lg-0.5.1.tar.gz

# Graph Neural Networks
torch-geometric>=2.3.0
dgl>=1.1.0
networkx>=3.0

# Knowledge Graph APIs
biopython>=1.81
pubchempy>=1.0.4
requests>=2.28.0
aiohttp>=3.8.0

# Configuration and Experiment Management
pyyaml>=6.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Visualization and Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
tensorboard>=2.12.0

# Utilities
tqdm>=4.65.0
colorlog>=6.7.0
python-dotenv>=1.0.0

# Statistical Analysis
statsmodels>=0.14.0
pingouin>=0.5.3

# Optional: Experiment Tracking
# wandb>=0.15.0

# Development
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0